// server/src/types/enums/enums.ts
export enum leadPriority {
  HIGH = "HIGH",
  LOW = "LOW",
  MEDIUM = "MEDIUM"  
}

export enum leadStatus {
  INITIATED = "INITIATED",
  CONSULTATIONLINKSENT = "CONSULTATIONLINKSENT",
  CONSULTATIONSCHEDULED = "CONSULTATIONSCHEDULED" ,
  CONSULTATIONDONE = "CONSULTATIONDONE" ,
  PAYMENTLINKSENT = "PAYMENTLINKSENT",
  PAYMENTDONE = "PAYMENTDONE" ,
  REJECTED = "REJECTED"
}